fibonacci function
a bit tricky. One thing I learned is that you don't have to declare the loop counter(a variable) at the beginning of a function. You just need to declare it with a type sign at the initialization part of the loop and that variable can be used only inside the loop.