a function for sorting strings
I spent a lot of time but still couldn't implement the functionality to read any number of strings below 10. So, in the end, I had to fix the number of strings to be read as 10.  Another issue is that I originally planned to use pointers for swapping the positions of strings, but I kept getting warnings during the compilation process. So, I had to resort to using the strncpy() function for swapping the positions of strings. I'm concerned that this might decrease the program's efficiency, but I don't know how to resolve this problem. I hope someone experienced can provide me with some insights on how to solve it.